refactor: 도메인별로 패키지 생성
