+++ src/main/java/com/coupang/marketplace/global/util/CryptoData.java
@@ -1,4 +1,4 @@
-package com.coupang.marketplace.util;
+package com.coupang.marketplace.global.util;
 
 import lombok.Builder;
 import lombok.Getter;
