+++ src/main/java/com/coupang/marketplace/global/util/Sha256Encryptor.java
@@ -1,4 +1,4 @@
-package com.coupang.marketplace.util;
+package com.coupang.marketplace.global.util;
 
 import org.springframework.stereotype.Component;
