--- src/main/java/com/coupang/marketplace/service/LoginService.java
@@ -1,9 +1,0 @@
-package com.coupang.marketplace.service;
-
-import com.coupang.marketplace.controller.SignInRequestDto;
-
-public interface LoginService {
-
-    void login(SignInRequestDto dto);
-
-}
