+++ src/main/java/com/coupang/marketplace/user/controller/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.coupang.marketplace.controller;
+package com.coupang.marketplace.user.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
